requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
azure-storage-queue
pandas
pydantic==2.7.2
extract_msg
#dotenv
pymupdf
llama-index
psutil
openai
tiktoken
poppler-utils
llama-index-llms-azure-openai
instructor
fuzzywuzzy
python-Levenshtein
usaddress
